reducer: Factor our separate loader
RepoLoader captures the repository package loading process separately
from the reducer and introduces an interface between them to make
testing the reducer easier.
